Navigate up a directory level to fix bug importing from segment_anything
Currently, if you run this script as specified in README.md, line 8 will throw ModuleNotFoundError: No module named 'segment_anything'. Currently, the only way to fix this is to either move amg.py up a directory, or add these two lines to it. This is because __init__.py is in the segment_anything dir,.only accessible from the main repo. This should also make relative file locations passed into the script be read to/written from with respect to the main directory, not scripts/.